[#9,#14,#17,irods/irods#7265] Minor reorganization + CMake TLC
- removed unused `main.cpp`
- reorganized project
- created object library target for `auth_check_wrapper.cpp`
- created interface library target for `ipam_client.hpp`
- quoted paths in cmake
- Removed unneeded `CMAKE_SHARED_LINKER_FLAGS_INIT` and `CMAKE_SHARED_LINKER_FLAGS_RELEASE_INIT`
- Added missing `CMAKE_EXE_LINKER_FLAGS_INIT` and `CMAKE_EXE_LINKER_FLAGS_RELEASE_INIT`
- Swapped quotes to angle brackets on irods includes
- Removed references to nonexistent include directory from cmake
- Removed unneeded `pam_conversation.hpp` include from `handshake_session.cpp`
- Bumped irods version and removed exact requirement
- Updated `CPACK_RPM_EXCLUDE_FROM_AUTO_FILELIST_ADDITION`
- Split into server and client packages (server package depends on client package)
- Replaced PAM `find_library` logic with a new find module
- Switched to `find_package` for fmt and spdlog
