+++ auth_check_wrapper/include/irods/private/pam/pam_auth_check_exception.hpp
