+++ plugin/include/irods/private/pam/handshake_session.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "ipam_client.hpp"
+#include "irods/private/pam/ipam_client.hpp"
 #include <utility>
 #include <condition_variable>
 #include <thread>
