--- src/main.cpp
@@ -1,4 +1,0 @@
-int main()
-{
-  return 0;
-}
